Create an abstract class to provide common interface for forces + move Forces to a separate .h and .cpp files
