+++ src/gebt_poc/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(${oturb_lib_name}
     PRIVATE
     dynamic_beam_element.cpp
     element.cpp
+    force.cpp
     gen_alpha_2D.cpp
     interpolation.cpp
     linearization_parameters.cpp
